8359108: Mac - When Swing starts First, native application menu doesn't work for JavaFX
